Move `BatchedSubmissionValidator` and spec into `batch` subpackage
